uses pytest parametrize to reduce code length
